% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temproal_beta.R
\name{beta_temporal_overlap}
\alias{beta_temporal_overlap}
\title{Compute the temporal change in beta diversity}
\usage{
beta_temporal_overlap(
  meta_composition_1,
  meta_composition_2,
  preprocess = T,
  ...
)
}
\arguments{
\item{meta_composition_1}{metacommunity at time 1}

\item{meta_composition_2}{metacommunity at time 2}

\item{preprocess}{whether to preprocess the data}

\item{...}{Arguments to be passed to methods}
}
\value{
the value of the geometric measure of beta_diversity
}
\description{
Compute the temporal change in beta diversity
}
