% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betavolume.R
\name{beta_volume}
\alias{beta_volume}
\title{Load a Matrix corresponds to metacommunity composition}
\usage{
beta_volume(meta_composition, weights = T)
}
\arguments{
\item{meta_composition}{metacommunity composition}

\item{weights}{whether the hyper points shouldbe weightedß}
}
\value{
the geometric measure of beta_diversity
}
\description{
This function loads a file as a matrix.
It assumes that columns correspond to local communities (plots or patches)
while the columns correspond to species.
The matrix entries correspond to species property (e.g, presence/absence).
}
