% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betavolume.R
\name{betavolume}
\alias{betavolume}
\title{Load a Matrix corresponds to metacommunity composition}
\usage{
betavolume(
  meta_composition,
  preprocess = T,
  weights = F,
  remove_unique = F,
  hypervolume_method,
  compressed_angle = 1,
  dim_threshold = 10,
  sparsity = F
)
}
\arguments{
\item{meta_composition}{metacommunity composition}

\item{preprocess}{Whether weights and remove_unique should be executed}

\item{weights}{whether non-unique compositions should be weighted}

\item{remove_unique}{whether unique compositions should be removed}

\item{hypervolume_method}{which method to estimate the hypervolume}

\item{compressed_angle}{how much the original hypervolume is changed by the similarity matrix}

\item{dim_threshold}{the threshold of which default hypervolume_method is used}

\item{sparsity}{whether the meta_composition is encoded as a sparse matrix}
}
\value{
the value of the geometric measure of beta_diversity
}
\description{
This function loads a file as a matrix.
It assumes that columns correspond to local communities (plots or patches)
while the columns correspond to species.
The matrix entries correspond to species property (e.g, presence/absence).
}
