% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disentangle_individual_contribution.R
\name{disentangle_individual_contribution}
\alias{disentangle_individual_contribution}
\title{Compute the patch-specific contribution to beta diversity}
\usage{
disentangle_individual_contribution(
  meta_composition,
  weights = F,
  remove_unique = F
)
}
\arguments{
\item{meta_composition}{metacommunity composition}

\item{weights}{whether non-unique compositions should be weighted}

\item{remove_unique}{whether unique compositions should be removed}
}
\value{
the value of the geometric measure of beta_diversity
}
\description{
Compute the patch-specific contribution to beta diversity
}
