% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disentangle_patch_contribution.R
\name{disentangle_patch_contribution}
\alias{disentangle_patch_contribution}
\title{Compute a metacommunity with}
\usage{
disentangle_patch_contribution()
}
\arguments{
\item{meta_composition}{metacommunity composition}

\item{weights}{whether non-unique compositions should be weighted}

\item{hypervolume_method}{which method to estimate the hypervolume}

\item{dim_threshold}{the threshold of which default hypervolume_method is used}

\item{remove_unique}{whether unique compositions should be removed}
}
\value{
the value of the geometric measure of beta_diversity
}
\description{
Compute a metacommunity with
}
