% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_meta_composition}
\alias{example_meta_composition}
\title{An example empirical dataset of metacommunity collected by A. H. Gentry and numerous additional collectors}
\format{
A matrix with 37 rows and 10 variables:
}
\source{
\url{http://doi.org/10.1126/science.1208584}
}
\usage{
example_meta_composition
}
\description{
An example empirical dataset of metacommunity collected by A. H. Gentry and numerous additional collectors
}
\keyword{datasets}
