% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compressed_angle.R
\name{get_compressed_angle}
\alias{get_compressed_angle}
\title{Calculate how much the original hypervolume is changed by the similarity matrix}
\usage{
get_compressed_angle(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{a matrix denote the level of species similarity or functional complementarity}
}
\value{
the size of the similarity angle
}
\description{
Calculate how much the original hypervolume is changed by the similarity matrix
}
