% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta_composition.R
\name{preprocess_meta_composition}
\alias{preprocess_meta_composition}
\title{Preprocess the metacommunity composition to facilitate computation of hypervolume}
\usage{
preprocess_meta_composition(meta_composition, weights = F, remove_unique = F)
}
\arguments{
\item{meta_composition}{metacommunity composition}

\item{weights}{whether non-unique compositions should be weighted}

\item{remove_unique}{whether unique compositions should be removed}
}
\value{
processed metacommunity composition
}
\description{
Preprocess the metacommunity composition to facilitate computation of hypervolume
}
