% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betavolume.R
\name{process_meta_composition}
\alias{process_meta_composition}
\title{Preprocess the metacommunity composition to facilitate computation of hypervolume}
\usage{
process_meta_composition(meta_composition, weights, remove_unique)
}
\arguments{
\item{meta_composition}{metacommunity composition}

\item{weights}{whether non-unique compositions should be weighted}

\item{remove_unique}{whether unique compositions should be removed}
}
\value{
the value of the geometric measure of beta_diversity
}
\description{
Preprocess the metacommunity composition to facilitate computation of hypervolume
}
