% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betavolume.R
\name{weight_composition}
\alias{weight_composition}
\title{Weight a metacommunity matrix according to the duplication scheme}
\usage{
weight_composition(meta_composition)
}
\arguments{
\item{meta_composition}{metacommunity composition}
}
\value{
the value of the geometric measure of beta_diversity
}
\description{
Weight a metacommunity matrix according to the duplication scheme
}
