% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betavolume.R
\name{betavolume}
\alias{betavolume}
\title{Load a Matrix corresponds to metacommunity composition}
\usage{
betavolume(
  meta_composition,
  weights = F,
  hypervolume_method,
  dim_threshold = 10,
  remove_unique = F
)
}
\arguments{
\item{meta_composition}{metacommunity composition}

\item{weights}{whether non-unique compositions should be weighted}

\item{hypervolume_method}{which method to estimate the hypervolume}

\item{dim_threshold}{the threshold of which default method is used}
}
\value{
the value of the geometric measure of beta_diversity
}
\description{
This function loads a file as a matrix.
It assumes that columns correspond to local communities (plots or patches)
while the columns correspond to species.
The matrix entries correspond to species property (e.g, presence/absence).
}
